+++ source/shared/loading/userdatavector.cpp
@@ -23,7 +23,7 @@
 QStringList UserDataVector::toStringList() const
 {
     QStringList list;
-    list.reserve(numValues());
+    list.reserve(static_cast<int>(numValues()));
 
     for(const auto& value : _values)
         list.append(value);
@@ -31,14 +31,14 @@ QStringList UserDataVector::toStringList() const
     return list;
 }
 
-int UserDataVector::numUniqueValues() const
+size_t UserDataVector::numUniqueValues() const
 {
     auto v = _values;
     std::sort(v.begin(), v.end());
     auto last = std::unique(v.begin(), v.end());
     v.erase(last, v.end());
 
-    return static_cast<int>(v.size());
+    return v.size();
 }
 
 bool UserDataVector::set(size_t index, const QString& newValue)
